updated backend folder and file structure
