+++ package.json
@@ -1,6 +1,7 @@
 {
   "devDependencies": {
     "express": "^4.19.2",
+    "express-async-handler": "^1.2.0",
     "moment": "^2.30.1"
   },
   "scripts": {
