+++ server/database/weatherCodes.js
@@ -1,0 +1,28 @@
+module.exports = {
+  0: "☀️",
+  1: "🌤️",
+  2: "⛅",
+  3: "☁️",
+  45: "🌫️",
+  48: "🌫️",
+  51: "💧",
+  56: "💧",
+  61: "💧",
+  66: "💧",
+  80: "💧",
+  53: "🌧️",
+  55: "🌧️",
+  57: "🌧️",
+  63: "🌧️",
+  65: "🌧️",
+  67: "🌧️",
+  81: "🌧️",
+  82: "🌧️",
+  71: "❄️",
+  77: "❄️",
+  85: "❄️",
+  73: "🌨️",
+  75: "🌨️",
+  86: "🌨️",
+  95: "⛈️",
+};
